==> Step 2 - Explore and Assess the Data.ipynb <==
import pandas as pd
import numpy as np

### Step 2: Explore and Assess the Data
#### Explore and Clean Data
- Identify data quality issues, like missing values, duplicate data, etc.
- Document steps necessary to clean the data

### Functions to use later

#function to describe dataframe from step 1
def stats_on_df(df, name):
    print("\nThere are {} rows and {} columns of data in the {} file".format(len(df), len(df.columns), name))
    print("columns are {}".format(df.columns))

#function to compare series to identify similarity
def matcher(series1, series2):
    matches = [i for i in series1 if i in series2]
    print(matches)

## Airports
Let's start with the **airport** data set found in /data_first_cleaning/airports.csv 

airport_df = pd.read_csv("data_first_cleaning/airports.csv")
stats_on_df(airport_df,"airports")

airport_df.airport_size.unique()

We will assume that only airports are relevant to immigration.

airports = ['small_airport', 'medium_airport', 'large_airport']
airport_df = airport_df[airport_df.airport_size.isin(airports)]
print("\nThere are {} rows of data in the airport file.".format(len(airport_df)))

print("continents = {}".format(airport_df.continent.unique())) #has nan
print("countries = {}".format(airport_df.country.unique())) #no nan

nans_cont = len(airport_df[airport_df.continent.isna()])/len(airport_df)
print("{}% of the continent data is nan".format(round(nans_cont*100,0)))

nan_countries = airport_df.country[airport_df.continent.isna()].unique()
notnan_countries = airport_df.country[airport_df.continent.notna()].unique()
matcher(nan_countries, notnan_countries)
#matches = [i for i in nan_countries if i in notnan_countries] #see if there is a country-based logic behind nans in continent data
#print(matches)

Since there is no intersection between these 2 arrays, we conclude that continent being nan is directly based on specific countries. Let's add the continent for the countries with nan. Data is cross-referenced with https://www.nationsonline.org/oneworld/country_code_list.htm

nan_countries

#north america countries in nan_countries array according to https://www.nationsonline.org
na_countries = ['US', 'PR', 'MX', 'BB', 'BS', 'BZ', 'CA', 'CR', 'CU', 'PA',
       'GT', 'HN', 'HT', 'JM', 'PM', 'TC', 'DO', 'NI', 'SV', 'KY', 'AG',
       'DM', 'GP', 'MQ', 'MF', 'BL', 'GD', 'VI', 'KN', 'LC', 'AW', 'AI', 'MS', 'TT', 'VG', 'VC', 'BM']
    
eu_countries = ['GL']

other_countries = ['BQ','CW', 'SX'] #will be dropped since they are not in the given source of country codes

airport_df.loc[airport_df['country'].isin(na_countries), 'continent'] = 'NA'
airport_df.loc[airport_df['country'].isin(eu_countries), 'continent'] = 'EU'
airport_df = airport_df[airport_df['continent'].notna()]
print("\nThere are {} rows of data in the airport file.".format(len(airport_df)))

We have now cleaned the continent column and only lost 5 rows of data due to potentially obsolete information.

print(str(len(airport_df[airport_df.airport_identifier==airport_df.gps_code])) + " rows with airport_identifier=gps_code.")
print(str(len(airport_df[airport_df.gps_code==airport_df.local_code])) + " rows with gps_code=local_code.")
print(str(len(airport_df[airport_df.airport_identifier==airport_df.local_code])) + " rows with airport_identifier=local_code.")

A lot of the data contains the same values for ident, gps_code, and local_code columns, but there is still a significant amount of data with different values in these columns therefore we will maintain all 3 columns columns.

airport_df.elevation_ft = pd.to_numeric(airport_df.elevation_ft, downcast='integer') #condense data
airport_df.head(2)
airport_df = airport_df[pd.notnull(airport_df.iata_code)]

airport_df.to_csv('data_second_cleaning/airports.csv', index=False) #save for to use in step 4

## Cities (Demographics)
Next we'll work with with the **cities** data set found in /data_first_cleaning/cities.csv 

cities_df = pd.read_csv("data_first_cleaning/cities.csv")
stats_on_df(cities_df,"cities")

cities_df.head()

cities_df.Race.unique()

For our purposes, we do not want separate rows for each city. We will reshape our data first to have 1 row per city. Then, since our immigration data does not give us a destination city but rather a destination state. We will condense the data into 1 row per state. 

#cities_melted = pd.melt(cities_df, id_vars=['Race'], value_vars=['Race Population'])
#cities_grouped = cities_df.groupby(['Race']).aggregate('sum')
cities_index = list(cities_df.columns)
cities_index.remove('Race')
cities_index.remove('Race Population')
cities_pivot = pd.pivot_table(cities_df, values='Race Population', index=cities_index, columns=['Race'], aggfunc=np.sum)
cities_pivot = cities_pivot.reset_index()

cities_pivot.State = cities_pivot['State Code']
cities_pivot.head()

states_pivot = cities_pivot.groupby(['State']).aggregate('sum')
cities_pivot2 = cities_pivot.loc[:,['State', 'Median Age', 'Average Household Size']]
states_pivot2 = cities_pivot2.groupby(['State']).aggregate('mean')
states_pivot.loc[:,['Median Age', 'Average Household Size']] = round(states_pivot2.loc[:,['Median Age', 'Average Household Size']],2)
sum(states_pivot['Total Population'])

Unfortunately, this data does not look to be complete as the US population is greater than 300 million. To address this, we will assume that the ratios are still valid and convert our data to percentages. We will convert total population to % of the entire country and the other population columns as % of the total for the specific state.

#convert population columns except total to % of the state
states_cols = list(states_pivot.columns)
states_cols.remove('Median Age')
states_cols.remove('Total Population')
states_cols.remove('Average Household Size')
for col in states_cols:
    states_pivot[col] = round(states_pivot[col]/states_pivot['Total Population'],3)*100
states_pivot.head()

#convert total population column to % of country total
states_pivot['Total Population'] = round(states_pivot['Total Population']/sum(states_pivot['Total Population']),3)*100

states_pivot.head()

states_pivot.to_csv('data_second_cleaning/states.csv', index=True) #save for to use in step 4

## Immigration Data
Let's explore the immigration sample data since it will be much easier to work with than the SAS data. This will allow us to identify the cleaning steps necessary on the larger set. 

imm_df = pd.read_csv("data_first_cleaning/immigration_sample.csv")
stats_on_df(imm_df,"immigration sample")

print(len(imm_df['count'].unique()))

There is no purpose for the count column as every row contains an identical value. This will not be maintained in our data model. 

print(imm_df.i94mode.unique())

Numeric coding will be changed from 1 to air, 2 to sea, and 3 to land respectively. Then the air data can be joined with the airports.csv data to provided additional information. 

## Temperature Data
Let's explore the temperature data and see what cleaning is necessary.

temp_df_full = pd.read_csv('../../data2/GlobalLandTemperaturesByCity.csv') 
stats_on_df(temp_df_full, 'temperatures all cities')

With over 8 million rows, this data will be very difficult to work with in Pandas. For our purposes, we do not need such extensive historical data. We can quickly truncate the data into a more manageable size by only maintaining recent years.

temp_df_full['Year'] = temp_df_full['dt'].astype(str).str[:4].astype(int)
temp_df_full.head()
temp_recent = temp_df_full[temp_df_full['Year']>1990]

stats_on_df(temp_recent, 'recent temps')

Now that our dataset is less than 1 million rows, we can use pandas much more efficiently. We would like to prep our data so that we can easy have the following fields via an SQL query When loading our data in the data pipeline:

- Highest monthly avg Temp (of most recent year)
- Lowest monthly avg Temp  (of most recent year)
- Temperature delta vs. 10 years ago (compare most recent year vs. 10 years prior)
- Temperature delta vs. 20 years ago (compare most recent year vs. 20 years prior)

Let's clean the data here so that we will be able to easily obtain the necessary fields in a SQL query in a later step. 

temp_recent.tail()

years = [1993, 2003, 2013]
temp_recent = temp_recent[temp_recent.Year.isin(years)]
stats_on_df(temp_recent, 'recent temps')

temp_recent[temp_recent['City']=="Detroit"].head(5)

Here we can see that the temperature data has city and country, however our immigration data only gives us the intended state and not city for destination. Therefore we'll need to reference our cities dataset to get state for each city.

citystate_df = pd.read_csv('data_first_cleaning/cities.csv')
stats_on_df(citystate_df, 'city state reference')

citystate_df = citystate_df[['City','State','Total Population']]
citystate_df = citystate_df.groupby(['City','State']).aggregate('max') #remove duplicates
citystate_df = citystate_df.reset_index()
citystate_df = citystate_df[['City','State']]
citystate_df[citystate_df.City=='Detroit']

temp_joined = temp_recent.set_index('City').join(citystate_df.set_index('City'))
temp_joined = temp_joined.reset_index()

stats_on_df(temp_joined, 'cities')

print(temp_joined.head())
temp_joined.loc[temp_joined.Country != 'United States','State'] = np.NaN
#df.loc[df.ID == 103, 'FirstName'] = "Matt"
temp_joined['Month'] = temp_joined['dt'].astype(str).str[5:7].astype(int)

del temp_joined['dt']

print(temp_joined[temp_joined.Country=='Mexico'].State.unique())

temp_df = temp_joined.copy()
#Create min max and average data
groupby_cols = ['Latitude','Longitude','Year', 'Country']
temp_df['avg_yrly_temp']=round(temp_df.groupby(groupby_cols)['AverageTemperature'].transform('mean'),2)
temp_df['mnthly_high_temp']=round(temp_df.groupby(groupby_cols)['AverageTemperature'].transform('max'),2)
temp_df['mnthly_low_temp']=round(temp_df.groupby(groupby_cols)['AverageTemperature'].transform('min'),2)
temp_df = temp_df.groupby(groupby_cols).last()
temp_df.reset_index(inplace=True)
print(temp_df.head(3))

print(temp_df.Country.unique())
print(len(temp_df))

#Collapse on year
temp_index = list(temp_df.columns)
temp_index.remove('avg_yrly_temp')
temp_index.remove('Year')
temp_df = temp_df.fillna('dummy') #otherwise it drops countries due to State=NaN
temp_pivot = pd.pivot_table(temp_df, values='avg_yrly_temp', index=temp_index, columns=['Year'], aggfunc=np.sum)
temp_pivot = temp_pivot.reset_index()
temp_pivot.columns = temp_pivot.columns.astype(str)
temp_pivot['State'][temp_pivot.State=='dummy'] = np.nan #revert to nan

#get delta values
temp_pivot[['1993','2003','2013']] = temp_pivot[['1993','2003','2013']].fillna(method='ffill')
temp_pivot = temp_pivot.iloc[2::3, :] #only keep 2013 rows
temp_pivot['temp_delta_10_yrs'] = temp_pivot['2013']-temp_pivot['2003']
temp_pivot['temp_delta_20_yrs'] = temp_pivot['2013']-temp_pivot['1993']
temp_pivot.reset_index()

#limit to relevant columns
temp_columns = ['City', 'Country', 'State','avg_yrly_temp', 'mnthly_high_temp', 'mnthly_low_temp', 'temp_delta_10_yrs', 'temp_delta_20_yrs', 'Latitude','Longitude']
temp_pivot = temp_pivot.rename(columns={'AverageTemperature':'avg_yrly_temp'})
temp_pivot = temp_pivot.rename_axis(None, axis = 1)
temp_pivot = temp_pivot.reset_index()
temp_pivot = temp_pivot[temp_columns]

temp_pivot.head(6)

temp_pivot.to_csv('data_second_cleaning/temperature.csv', index=False) #save for to use in step 4

This new temperature dataset is less than 1MB whereas we started with a dataset >500MB.
